Forgot to move a domain class to the domain package
