+++ src/com/movierentals/RentalInfo.java
@@ -2,6 +2,7 @@ package com.movierentals;
 
 
 import com.movierentals.domain.Customer;
+import com.movierentals.domain.MovieRental;
 import com.movierentals.services.MovieRepository;
 
 import static com.movierentals.domain.MovieCategory.*;
