+++ src/com/movierentals/domain/MovieRental.java
@@ -1,4 +1,4 @@
-package com.movierentals;
+package com.movierentals.domain;
 
 
 public class MovieRental {
